Update BaselineStatusResults.ojo
Use callback mechanism to trigger the pie chart being drawn instead of having it execute on the next line.

the ajax callback mechanism will send the data as a parameter to `drawpiechart` which i'm not currently using. This might be a better option to get the chart data instead of getting it from the redrawn table directly, though this data would have to be transformed in a similar fashion.

Changed the empty table messaging slightly.